initial setup API and serializers
