+++ app/controllers/api/samples_controller.rb
