+++ app/serializers/camera_serializer.rb
@@ -1,0 +1,3 @@
+class CameraSerializer < ActiveModel::Serializer
+  attributes :brand, :model, :price, :product_photo, :description
+end
